+++ client/pages/login/index.tsx
@@ -1,12 +1,12 @@
 import styled from 'styled-components';
-import apiUrl from '@constants/apiUrl';
 import ComponentInfoContext from '@utils/context/ComponentInfoContext';
 import { componentType, contentType } from '@constants/identifier';
-import useClickEventLog from '@hooks/useClickEventLog';
 import { Cookies } from 'react-cookie';
 import { useEffect } from 'react';
 import ComponentInfoWrapper from '@utils/context/ComponentInfoWrapper';
 import Button from '@components/atoms/Button';
+import NaverLoginButton from '@components/atoms/NaverLoginButton';
+import LoginForm from '@components/molecules/LoginForm';
 import Link from 'next/link';
 import HeadsetIcon from '@material-ui/icons/Headset';
 
@@ -26,60 +26,7 @@ const LoginContainer = styled.div`
     padding-bottom: 200px;
 `;
 
-const Form = styled.form`
-    display: flex;
-    flex-flow: column;
-    align-items: center;
-    justify-content: center;
-`;
-
-const LoginInput = styled.input`
-    background-color: transperant;
-    display: block;
-    margin: 0 auto;
-    width: 231.19px;
-    height: 40px;
-    margin-bottom: 10px;
-    &:focus {
-        outline: none;
-    }
-`;
-
-const LoginButton = styled.input`
-    focus: none;
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-    padding: 9px 15px;
-    color: black;
-    border: 1px solid #d7d7d7;
-    background-color: white;
-    width: 231.19px;
-    height: 50px;
-    border-radius: 4px;
-    box-shadow: none;
-    cursor: pointer;
-    font-size: 15px;
-    line-height: 1;
-    outline: none;
-    -webkit-text-decoration: none;
-    text-decoration: none;
-    white-space: nowrap;
-`;
-
-const NaverLogin = styled.a`
-    display: block;
-    margin-top: 40px;
-`;
-
-const Title = styled.p`
-    font-size: 20px;
-    font-weight: 700;
-`;
-
 const Login = () => {
-    const localClick = useClickEventLog({ userId: null, href: `${apiUrl.login}/local-login` });
-    const naverClick = useClickEventLog({ userId: null, href: `${apiUrl.login}/naver-login` });
     const cookies = new Cookies();
     useEffect(() => {
         if (cookies.get('fail')) {
@@ -92,21 +39,10 @@ const Login = () => {
         <ComponentInfoContext.Provider value={{ componentId: componentType.loginForm }}>
             <LoginContainer>
                 <ComponentInfoWrapper componentId={contentType.localLogin}>
-                    <Form action={`${apiUrl.login}/local-login`} method="post">
-                        <Title>miniVIBE 로그인</Title>
-                        <LoginInput type="email" name="email" placeholder="이메일을 입력해주세요" />
-                        <LoginInput id="pw" type="password" name="pw" placeholder="비밀번호를 입력해주세요" />
-                        <LoginButton id="local-login" type="submit" value="로그인" onClick={localClick} />
-                    </Form>
+                    <LoginForm />
                 </ComponentInfoWrapper>
                 <ComponentInfoWrapper componentId={contentType.naverLogin}>
-                    <NaverLogin id="naver-login" href={`${apiUrl.login}/naver-login`} onClick={naverClick}>
-                        <img
-                            height="50"
-                            src="https://user-images.githubusercontent.com/60503734/102323140-717b2300-3fc3-11eb-886d-e124e75d3169.PNG"
-                            alt="네이버로 로그인"
-                        />
-                    </NaverLogin>
+                    <NaverLoginButton />
                 </ComponentInfoWrapper>
                 <ComponentInfoWrapper componentId={contentType.join}>
                     <Link href="/join">
